requirements.txt
streamlit>=1.36
llama-cpp-python>=0.2.78
pandas>=2.2
python-dotenv>=1.0
